Reduced the duplicated code.
